fix(channel): don't push async handled commands to stream
As there's no piping mechanism between channels and consumers, channel's
readable buffer gets filled with deliveries and the stream will stop
once the buffer reaches the high watermark point.

There may be a way for this to be written in a more idiomatic way; I've
yet to figure out how.