fix(frame_decoder): call extractFrames from where it left last time
If we have a residual chunk when extractFrames is called, and the new
chunk has more bytes than the incomplete frame needs, first we complete
the frame, and then call extractFrames with the remaining bytes of the
chunk. Calculation of the remaining bytes in the buffer was previously
faulty, and is fixed by this commit.